pyproject.toml
[project]
name = "ical2jcal"
dynamic = ["version"]
description = "Convert between iCalendar and jCalendar formats."
authors = [
    { name = "Nicco Kunzmann", email = "niccokunzmann@rambler.ru" }]
license = "GPL-3.0-or-later"
license-files = ["LICENSE*"]
readme = "README.md"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Free Threading :: 3 - Stable",
    "Typing :: Typed",
    "Topic :: File Formats :: JSON",
    "Topic :: File Formats",
    "Topic :: Communications",
    "Topic :: Utilities",
    # https://pypi.org/classifiers/
]
requires-python = ">=3.10"
dependencies = [
    "rich",
    "typer-slim[standard]",
    "icalendar>=7.0.0a2",
]

[project.urls]
Homepage = "https://github.com/niccokunzmann/ical2jcal/"
Documentation = "https://github.com/niccokunzmann/ical2jcal/"
Repository = "https://github.com/niccokunzmann/ical2jcal/"

[project.scripts]
ical2jcal = "ical2jcal.cli:ical2jcal"
jcal2ical = "ical2jcal.cli:jcal2ical"

[dependency-groups]
nox = [
    "nox-uv",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-randomly",
]
type_check = [
    "mypy",
    # Add "types-" stub packages as needed: https://github.com/python/typeshed/tree/main/stubs
]
lint = [
    "ruff",
]
docs = [
    "mkdocs-material",
    "mkdocs-htmlproofer-plugin",

    # Python API documentation (not neccessary for applications).
    "mkdocstrings[python]",
    # Autodoc.
    "mkdocs-api-autonav",
]
licenses = [
    "pip-licenses-cli",
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
fallback-version = "0.0.0"

[tool.uv]
default-groups = "all"
# Non-package applications can uncomment the following line and delete [build-system] section to
# avoid being built or installed: https://docs.astral.sh/uv/reference/settings/#package
#package = false

[tool.mypy]
ignore_missing_imports = true
strict = true
enable_error_code = ["deprecated", "exhaustive-match", "explicit-override"]
# If certain strict config options are too pedantic for a project,
# disable them selectively here by setting to false.

[tool.ruff]
line-length = 99
src = ["src"]
# Ruff will automatically exclude all files listed in .gitignore as well as common temporary Python
# tool directories.
# To exclude additional folders, use extend-exclude.

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "RUF", # ruff
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "PT", # flake-pytest-style
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
extend-ignore = [
    "RUF005",
    "RUF012",
]
unfixable = [
    # Disable removing unused imports by default and only enable within nox so editors don't delete
    # unused imports while the user is in the middle of editing a file on save.
    "F401",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false
# For non-src directory projects, explicitly set top level package names:
# known-first-party = ["my-app"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
]
xfail_strict = true
filterwarnings = [
    # When running tests, treat warnings as errors (e.g. -Werror).
    # See: https://docs.pytest.org/en/latest/reference/reference.html#confval-filterwarnings
    "error",
    # Add additional warning suppressions as needed here. For example, if a third-party library
    # is throwing a deprecation warning that needs to be fixed upstream:
    # "ignore::DeprecationWarning:typer",
]

[tool.coverage.run]
branch = true

# To globally exclude additional code blocks from code coverage reporting, see:
# https://coverage.readthedocs.io/en/latest/excluding.html#advanced-exclusion
[tool.coverage.report]
exclude_also = [
    "case _ as unreachable",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
